Refactor state loading and genesis doc provider into state package
